+++ bin/bench-tx/src/main.rs
@@ -98,11 +98,10 @@ pub fn benchmark_p2id() -> anyhow::Result<TransactionMeasurements> {
     .unwrap();
 
     let tx_script_target =
-        TransactionScript::compile(DEFAULT_AUTH_SCRIPT, [], TransactionKernel::assembler())
-            .unwrap();
+        TransactionScript::compile(DEFAULT_AUTH_SCRIPT, TransactionKernel::assembler()).unwrap();
 
     let tx_context = TransactionContextBuilder::new(target_account.clone())
-        .input_notes(vec![note])
+        .extend_input_notes(vec![note])
         .tx_script(tx_script_target)
         .authenticator(Some(falcon_auth))
         .build();