+++ crates/miden-objects/src/transaction/tx_args.rs
@@ -1,6 +1,5 @@
 use alloc::{
     collections::{BTreeMap, BTreeSet},
-    string::ToString,
     sync::Arc,
     vec::Vec,
 };
@@ -10,27 +9,33 @@ use miden_crypto::merkle::InnerNodeInfo;
 
 use super::{AccountInputs, Digest, Felt, Word};
 use crate::{
-    Hasher, MastForest, MastNodeId, TransactionScriptError,
+    EMPTY_WORD, MastForest, MastNodeId, TransactionScriptError,
     note::{NoteId, NoteRecipient},
     utils::serde::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable},
     vm::{AdviceInputs, AdviceMap, Program},
 };
 
-// TRANSACTION ARGS
+// TRANSACTION ARGUMENTS
 // ================================================================================================
 
 /// Optional transaction arguments.
 ///
 /// - Transaction script: a program that is executed in a transaction after all input notes scripts
 ///   have been executed.
+/// - Transaction script argument: a [`Word`], which will be pushed to the operand stack before the
+///   transaction script execution. If this argument is not specified, the [`EMPTY_WORD`] would be
+///   used as a default value. If the [AdviceInputs] are propagated with some user defined map
+///   entires, this script argument could be used as a key to access the corresponding value.
 /// - Note arguments: data put onto the stack right before a note script is executed. These are
 ///   different from note inputs, as the user executing the transaction can specify arbitrary note
 ///   args.
-/// - Advice inputs: Provides data needed by the runtime, like the details of public output notes.
-/// - Account inputs: Provides account data that will be accessed in the transaction.
+/// - Advice inputs: provides data needed by the runtime, like the details of public output notes.
+/// - Foreign account inputs: provides foreign account data that will be used during the foreign
+///   procedure invocation (FPI).
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
 pub struct TransactionArgs {
     tx_script: Option<TransactionScript>,
+    tx_script_arg: Word,
     note_args: BTreeMap<NoteId, Word>,
     advice_inputs: AdviceInputs,
     foreign_account_inputs: Vec<AccountInputs>,
@@ -40,40 +45,48 @@ impl TransactionArgs {
     // CONSTRUCTORS
     // --------------------------------------------------------------------------------------------
 
-    /// Returns new [TransactionArgs] instantiated with the provided transaction script and note
-    /// arguments.
-    ///
-    /// If tx_script is provided, this also adds all mappings from the transaction script inputs
-    /// to the advice inputs' map.
-    pub fn new(
-        tx_script: Option<TransactionScript>,
-        note_args: Option<BTreeMap<NoteId, Word>>,
-        advice_map: AdviceMap,
-        foreign_account_inputs: Vec<AccountInputs>,
-    ) -> Self {
-        let mut advice_inputs = AdviceInputs::default().with_map(advice_map);
-        // add transaction script inputs to the advice inputs' map
-        if let Some(ref tx_script) = tx_script {
-            advice_inputs
-                .extend_map(tx_script.inputs().iter().map(|(hash, input)| (*hash, input.clone())))
-        }
-
+    /// Returns new [TransactionArgs] instantiated with the provided transaction script, advice
+    /// map and foreign account inputs.
+    pub fn new(advice_map: AdviceMap, foreign_account_inputs: Vec<AccountInputs>) -> Self {
         Self {
-            tx_script,
-            note_args: note_args.unwrap_or_default(),
-            advice_inputs,
+            tx_script: None,
+            tx_script_arg: EMPTY_WORD,
+            note_args: Default::default(),
+            advice_inputs: AdviceInputs::default().with_map(advice_map),
             foreign_account_inputs,
         }
     }
 
     /// Returns new [TransactionArgs] instantiated with the provided transaction script.
+    ///
+    /// If the transaction script is already set, it will be overwritten with the newly provided
+    /// one.
     #[must_use]
     pub fn with_tx_script(mut self, tx_script: TransactionScript) -> Self {
         self.tx_script = Some(tx_script);
         self
     }
 
+    /// Returns new [TransactionArgs] instantiated with the provided transaction script and its
+    /// argument.
+    ///
+    /// If the transaction script and argument are already set, they will be overwritten with the
+    /// newly provided ones.
+    #[must_use]
+    pub fn with_tx_script_and_arg(
+        mut self,
+        tx_script: TransactionScript,
+        tx_script_arg: Word,
+    ) -> Self {
+        self.tx_script = Some(tx_script);
+        self.tx_script_arg = tx_script_arg;
+        self
+    }
+
     /// Returns new [TransactionArgs] instantiated with the provided note arguments.
+    ///
+    /// If the note arguments were already set, they will be overwritten with the newly provided
+    /// ones.
     #[must_use]
     pub fn with_note_args(mut self, note_args: BTreeMap<NoteId, Word>) -> Self {
         self.note_args = note_args;
@@ -88,17 +101,28 @@ impl TransactionArgs {
         self.tx_script.as_ref()
     }
 
+    /// Returns the transaction script argument, or [`EMPTY_WORD`] if the argument was not
+    /// specified.
+    ///
+    /// This argument could be potentially used as a key to access the advice map during the
+    /// transaction script execution. Notice that the corresponding map entry should be provided
+    /// separately during the creation with the [`TransactionArgs::new`] or using the
+    /// [`TransactionArgs::extend_advice_map`] method.
+    pub fn tx_script_arg(&self) -> Word {
+        self.tx_script_arg
+    }
+
     /// Returns a reference to a specific note argument.
     pub fn get_note_args(&self, note_id: NoteId) -> Option<&Word> {
         self.note_args.get(&note_id)
     }
 
-    /// Returns a reference to the args [AdviceInputs].
+    /// Returns a reference to the internal [AdviceInputs].
     pub fn advice_inputs(&self) -> &AdviceInputs {
         &self.advice_inputs
     }
 
-    /// Returns a reference to the foreign account inputs in the transaction args.
+    /// Returns a reference to the foreign account inputs in the transaction arguments.
     pub fn foreign_account_inputs(&self) -> &[AccountInputs] {
         &self.foreign_account_inputs
     }
@@ -172,6 +196,7 @@ impl TransactionArgs {
 impl Serializable for TransactionArgs {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
         self.tx_script.write_into(target);
+        self.tx_script_arg.write_into(target);
         self.note_args.write_into(target);
         self.advice_inputs.write_into(target);
         self.foreign_account_inputs.write_into(target);
@@ -181,12 +206,14 @@ impl Serializable for TransactionArgs {
 impl Deserializable for TransactionArgs {
     fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
         let tx_script = Option::<TransactionScript>::read_from(source)?;
+        let tx_script_arg = Word::read_from(source)?;
         let note_args = BTreeMap::<NoteId, Word>::read_from(source)?;
         let advice_inputs = AdviceInputs::read_from(source)?;
         let foreign_account_inputs = Vec::<AccountInputs>::read_from(source)?;
 
         Ok(Self {
             tx_script,
+            tx_script_arg,
             note_args,
             advice_inputs,
             foreign_account_inputs,
@@ -202,109 +229,46 @@ impl Deserializable for TransactionArgs {
 /// A transaction script is a program that is executed in a transaction after all input notes
 /// have been executed.
 ///
-/// The [TransactionScript] object is composed of:
-/// - An executable program defined by a [MastForest] and an associated entrypoint.
-/// - A set of transaction script inputs defined by a map of key-value inputs that are loaded into
-///   the advice inputs' map such that the transaction script can access them.
-/// - A script arguments key defined as an optional [`Digest`]: if present, this key will be pushed
-///   to the operand stack before the transaction script execution and could be used to get the
-///   script arguments array. See [`TransactionScript::with_args`] for more details.
+/// The [TransactionScript] object is composed of an executable program defined by a [MastForest]
+/// and an associated entrypoint.
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct TransactionScript {
     mast: Arc<MastForest>,
     entrypoint: MastNodeId,
-    inputs: BTreeMap<Digest, Vec<Felt>>,
-    args_key: Option<Digest>,
 }
 
 impl TransactionScript {
     // CONSTRUCTORS
     // --------------------------------------------------------------------------------------------
 
-    /// Returns a new [TransactionScript] instantiated with the provided code and inputs.
-    pub fn new(code: Program, inputs: impl IntoIterator<Item = (Word, Vec<Felt>)>) -> Self {
-        Self {
-            entrypoint: code.entrypoint(),
-            mast: code.mast_forest().clone(),
-            inputs: inputs.into_iter().map(|(k, v)| (k.into(), v)).collect(),
-            args_key: None,
-        }
+    /// Returns a new [TransactionScript] instantiated with the provided code.
+    pub fn new(code: Program) -> Self {
+        Self::from_parts(code.mast_forest().clone(), code.entrypoint())
     }
 
-    /// Returns a new [TransactionScript] compiled from the provided source code and inputs using
-    /// the specified assembler.
+    /// Returns a new [TransactionScript] compiled from the provided source code using the specified
+    /// assembler.
     ///
     /// # Errors
     /// Returns an error if the compilation of the provided source code fails.
     pub fn compile(
         source_code: impl Compile,
-        inputs: impl IntoIterator<Item = (Word, Vec<Felt>)>,
         assembler: Assembler,
     ) -> Result<Self, TransactionScriptError> {
         let program = assembler
             .assemble_program(source_code)
             .map_err(TransactionScriptError::AssemblyError)?;
-        Ok(Self::new(program, inputs))
+        Ok(Self::new(program))
     }
 
-    /// Returns a new [TransactionScript] instantiated from the provided components.
-    ///
-    /// If the `args_key` is present, it's expected that `inputs` map will have it as a key, with
-    /// the value being equal to the provided script arguments array. See
-    /// [`TransactionScript::with_args`] for more details.
-    ///
-    /// # Errors
-    /// Returns an error if:
-    /// - The provided `args_key` is not presented in the `inputs` map.
+    /// Returns a new [TransactionScript] instantiated from the provided MAST forest and entrypoint.
     ///
     /// # Panics
     /// Panics if the specified entrypoint is not in the provided MAST forest.
-    pub fn from_parts(
-        mast: Arc<MastForest>,
-        entrypoint: MastNodeId,
-        inputs: BTreeMap<Digest, Vec<Felt>>,
-        args_key: Option<Digest>,
-    ) -> Result<Self, TransactionScriptError> {
+    pub fn from_parts(mast: Arc<MastForest>, entrypoint: MastNodeId) -> Self {
         assert!(mast.get_node_by_id(entrypoint).is_some());
 
-        // check that provided `args_key` is presented in the `inputs` map
-        if let Some(args_key) = args_key {
-            if !inputs.contains_key(&args_key) {
-                return Err(TransactionScriptError::MissingScriptArgsKeyEntry(args_key));
-            }
-        }
-
-        Ok(Self { mast, entrypoint, inputs, args_key })
-    }
-
-    // MUTATORS
-    // --------------------------------------------------------------------------------------------
-
-    /// Sets the `args_key` to the commitment of the provided arguments slice and extends the
-    /// `inputs` map with the `COMPUTED_COMMITMENT -> [[script_args]]` entry.
-    ///
-    /// Script arguments is an optional array of [`Felt`]s which could be easily accessed at the
-    /// beginning of the transaction script execution. The commitment of this array (`args_key`) is
-    /// automatically pushed to the operand stack at the beginning of the transaction script
-    /// execution and the underlying arguments can be accessed using the `adv.push_mapval`
-    /// and `adv_push.n` instructions.
-    pub fn with_args(mut self, script_args: &[Felt]) -> Result<Self, TransactionScriptError> {
-        let args_key = Hasher::hash_elements(script_args);
-        let old_map_value = self.inputs.insert(args_key, script_args.to_vec());
-
-        // check that a new map entry will not overwrite an existing one
-        if let Some(old_value) = old_map_value {
-            if old_value != script_args {
-                return Err(TransactionScriptError::ScriptArgsCollision {
-                    key: args_key,
-                    new_value: script_args.to_vec(),
-                    old_value,
-                });
-            }
-        }
-
-        self.args_key = Some(args_key);
-        Ok(self)
+        Self { mast, entrypoint }
     }
 
     // PUBLIC ACCESSORS
@@ -319,17 +283,6 @@ impl TransactionScript {
     pub fn root(&self) -> Digest {
         self.mast[self.entrypoint].digest()
     }
-
-    /// Returns a reference to the inputs for this transaction script.
-    pub fn inputs(&self) -> &BTreeMap<Digest, Vec<Felt>> {
-        &self.inputs
-    }
-
-    /// Returns the commitment of the transaction script arguments, or [`None`] if they were not
-    /// specified.
-    pub fn args_key(&self) -> Option<Digest> {
-        self.args_key
-    }
 }
 
 // SERIALIZATION
@@ -339,8 +292,6 @@ impl Serializable for TransactionScript {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
         self.mast.write_into(target);
         target.write_u32(self.entrypoint.as_u32());
-        self.inputs.write_into(target);
-        self.args_key.write_into(target);
     }
 }
 
@@ -348,11 +299,8 @@ impl Deserializable for TransactionScript {
     fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
         let mast = MastForest::read_from(source)?;
         let entrypoint = MastNodeId::from_u32_safe(source.read_u32()?, &mast)?;
-        let inputs = BTreeMap::<Digest, Vec<Felt>>::read_from(source)?;
-        let script_args_key = Option::<Digest>::read_from(source)?;
 
-        Self::from_parts(Arc::new(mast), entrypoint, inputs, script_args_key)
-            .map_err(|e| DeserializationError::InvalidValue(e.to_string()))
+        Ok(Self::from_parts(Arc::new(mast), entrypoint))
     }
 }
 
@@ -367,10 +315,10 @@ mod tests {
 
     #[test]
     fn test_tx_args_serialization() {
-        let args = TransactionArgs::new(None, None, AdviceMap::default(), std::vec::Vec::default());
-        let bytes: std::vec::Vec<u8> = args.to_bytes();
+        let tx_args = TransactionArgs::new(AdviceMap::default(), std::vec::Vec::default());
+        let bytes: std::vec::Vec<u8> = tx_args.to_bytes();
         let decoded = TransactionArgs::read_from_bytes(&bytes).unwrap();
 
-        assert_eq!(args, decoded);
+        assert_eq!(tx_args, decoded);
     }
 }