+++ crates/miden-objects/src/transaction/tx_witness.rs
@@ -13,8 +13,9 @@ use crate::utils::serde::{ByteReader, Deserializable, DeserializationError, Seri
 /// A transaction witness consists of:
 /// - Transaction inputs which contain information about the initial state of the account, input
 ///   notes, block header etc.
-/// - Optional transaction arguments which may contain a transaction script, note arguments, and any
-///   additional advice data to initialize the advice provide with prior to transaction execution.
+/// - Optional transaction arguments which may contain a transaction script, note arguments,
+///   transaction script argument and any additional advice data to initialize the advice provider
+///   with prior to transaction execution.
 /// - Advice witness which contains all data requested by the VM from the advice provider while
 ///   executing the transaction program.
 ///