+++ crates/miden-testing/src/mock_chain/chain.rs
@@ -587,7 +587,6 @@ impl MockChain {
         if mock_account.authenticator().is_some() {
             let tx_script = TransactionScript::compile(
                 DEFAULT_AUTH_SCRIPT,
-                vec![],
                 TransactionKernel::testing_assembler_with_mock_account(),
             )
             .unwrap();