+++ crates/miden-testing/src/tx_context/builder.rs
@@ -5,7 +5,7 @@ use alloc::{collections::BTreeMap, vec::Vec};
 
 use miden_lib::{transaction::TransactionKernel, utils::word_to_masm_push_string};
 use miden_objects::{
-    FieldElement,
+    EMPTY_WORD, FieldElement,
     account::{Account, AccountId},
     assembly::Assembler,
     asset::{Asset, FungibleAsset, NonFungibleAsset},
@@ -78,6 +78,7 @@ pub struct TransactionContextBuilder {
     foreign_account_inputs: Vec<AccountInputs>,
     input_notes: Vec<Note>,
     tx_script: Option<TransactionScript>,
+    tx_script_arg: Word,
     note_args: BTreeMap<NoteId, Word>,
     transaction_inputs: Option<TransactionInputs>,
     rng: ChaCha20Rng,
@@ -93,6 +94,7 @@ impl TransactionContextBuilder {
             expected_output_notes: Vec::new(),
             rng: ChaCha20Rng::from_seed([0_u8; 32]),
             tx_script: None,
+            tx_script_arg: EMPTY_WORD,
             authenticator: None,
             advice_inputs: Default::default(),
             transaction_inputs: None,
@@ -122,6 +124,7 @@ impl TransactionContextBuilder {
             advice_inputs: Default::default(),
             rng: ChaCha20Rng::from_seed([0_u8; 32]),
             tx_script: None,
+            tx_script_arg: EMPTY_WORD,
             transaction_inputs: None,
             note_args: BTreeMap::new(),
             foreign_account_inputs: vec![],
@@ -158,9 +161,19 @@ impl TransactionContextBuilder {
         self
     }
 
-    /// Override and set the [AdviceInputs]
-    pub fn advice_inputs(mut self, advice_inputs: AdviceInputs) -> Self {
-        self.advice_inputs = advice_inputs;
+    /// Extend the advice inputs with the provided [AdviceInputs] instance.
+    pub fn extend_advice_inputs(mut self, advice_inputs: AdviceInputs) -> Self {
+        self.advice_inputs.extend(advice_inputs);
+        self
+    }
+
+    /// Extend the advice inputs map with the provided iterator.
+    pub fn extend_advice_map(
+        mut self,
+        map_entries: impl IntoIterator<Item = (Word, Vec<Felt>)>,
+    ) -> Self {
+        self.advice_inputs
+            .extend_map(map_entries.into_iter().map(|(hash, input)| (hash.into(), input)));
         self
     }
 
@@ -177,14 +190,8 @@ impl TransactionContextBuilder {
     }
 
     /// Extend the set of used input notes
-    pub fn input_notes(mut self, input_notes: Vec<Note>) -> Self {
+    pub fn extend_input_notes(mut self, input_notes: Vec<Note>) -> Self {
         self.input_notes.extend(input_notes);
-        self
-    }
-
-    // Set the desired note args
-    pub fn note_args(mut self, args: BTreeMap<NoteId, Word>) -> Self {
-        self.note_args = args;
         self
     }
 
@@ -194,14 +201,26 @@ impl TransactionContextBuilder {
         self
     }
 
+    /// Set the transaction script argument
+    pub fn tx_script_arg(mut self, tx_script_arg: Word) -> Self {
+        self.tx_script_arg = tx_script_arg;
+        self
+    }
+
     /// Set the desired transaction inputs
     pub fn tx_inputs(mut self, tx_inputs: TransactionInputs) -> Self {
         self.transaction_inputs = Some(tx_inputs);
         self
     }
 
-    /// Defines the expected output notes
-    pub fn expected_notes(mut self, output_notes: Vec<OutputNote>) -> Self {
+    /// Extend the note arguments map with the provided one.
+    pub fn extend_note_args(mut self, note_args: BTreeMap<NoteId, Word>) -> Self {
+        self.note_args.extend(note_args);
+        self
+    }
+
+    /// Extend the expected output notes.
+    pub fn extend_expected_output_notes(mut self, output_notes: Vec<OutputNote>) -> Self {
         let output_notes = output_notes.into_iter().filter_map(|n| match n {
             OutputNote::Full(note) => Some(note),
             OutputNote::Partial(_) => None,
@@ -465,7 +484,7 @@ impl TransactionContextBuilder {
             fungible_asset_1,
         );
 
-        self.input_notes(vec![input_note1])
+        self.extend_input_notes(vec![input_note1])
     }
 
     /// Adds a set of input notes that output notes in an asset-preserving manner.
@@ -504,7 +523,7 @@ impl TransactionContextBuilder {
             &output_note2,
         );
 
-        self.input_notes(vec![input_note1, input_note2])
+        self.extend_input_notes(vec![input_note1, input_note2])
     }
 
     pub fn with_mock_notes_preserved_with_account_vault_delta(mut self) -> Self {
@@ -546,7 +565,7 @@ impl TransactionContextBuilder {
         let input_note5 = self
             .input_note_transfer(sender, [fungible_asset_1, fungible_asset_3, nonfungible_asset_1]);
 
-        self.input_notes(vec![input_note1, input_note2, input_note5])
+        self.extend_input_notes(vec![input_note1, input_note2, input_note5])
     }
 
     pub fn with_mock_notes_too_many_fungible_input(mut self) -> Self {
@@ -586,7 +605,7 @@ impl TransactionContextBuilder {
         let input_note3 =
             self.input_note_simple(sender, [fungible_asset_2, fungible_asset_3], [2u32.into()]);
 
-        self.input_notes(vec![input_note1, input_note2, input_note3])
+        self.extend_input_notes(vec![input_note1, input_note2, input_note3])
     }
 
     pub fn with_mock_notes_too_many_non_fungible_input(mut self) -> Self {
@@ -626,7 +645,7 @@ impl TransactionContextBuilder {
         );
         let input_note4 = self.input_note_simple(sender, [nonfungible_asset_1], [1u32.into()]);
 
-        self.input_notes(vec![input_note1, input_note2, input_note4])
+        self.extend_input_notes(vec![input_note1, input_note2, input_note4])
     }
 
     /// Builds the [TransactionContext].
@@ -662,12 +681,14 @@ impl TransactionContextBuilder {
             },
         };
 
-        let mut tx_args = TransactionArgs::new(
-            self.tx_script,
-            Some(self.note_args),
-            AdviceMap::default(),
-            self.foreign_account_inputs,
-        );
+        let tx_args = TransactionArgs::new(AdviceMap::default(), self.foreign_account_inputs)
+            .with_note_args(self.note_args);
+
+        let mut tx_args = if let Some(tx_script) = self.tx_script {
+            tx_args.with_tx_script_and_arg(tx_script, self.tx_script_arg)
+        } else {
+            tx_args
+        };
 
         tx_args.extend_advice_inputs(self.advice_inputs.clone());
         tx_args.extend_output_note_recipients(self.expected_output_notes.clone());