+++ crates/miden-testing/tests/integration/scripts/faucet.rs
@@ -69,8 +69,7 @@ fn prove_faucet_contract_mint_fungible_asset_succeeds() {
     );
 
     let tx_script =
-        TransactionScript::compile(tx_script_code, vec![], TransactionKernel::testing_assembler())
-            .unwrap();
+        TransactionScript::compile(tx_script_code, TransactionKernel::testing_assembler()).unwrap();
     let tx_context = mock_chain
         .build_tx_context(faucet.account().id(), &[], &[])
         .tx_script(tx_script)
@@ -138,8 +137,7 @@ fn faucet_contract_mint_fungible_asset_fails_exceeds_max_supply() {
     );
 
     let tx_script =
-        TransactionScript::compile(tx_script_code, vec![], TransactionKernel::testing_assembler())
-            .unwrap();
+        TransactionScript::compile(tx_script_code, TransactionKernel::testing_assembler()).unwrap();
     let tx = mock_chain
         .build_tx_context(faucet.account().id(), &[], &[])
         .tx_script(tx_script)